% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{make_letters}
\alias{make_letters}
\title{Build a sequence of letters to extend past 26 facets, in the style of
spreadsheet column indices (Y, Z, AA, AB, AC, ...).}
\usage{
make_letters(y)
}
\arguments{
\item{y}{A vector of indices to create letters for}
}
\value{
A vector of letters potentially extending past a–z
}
\description{
(Adapted from cellranger::letter_to_num)
}
