% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{numbering_present}
\alias{numbering_present}
\title{Detect the use of numbering variables}
\usage{
numbering_present(template)
}
\arguments{
\item{template}{The string template to check}
}
\value{
A boolean: TRUE if a numbering variable is present.
}
\description{
Numbering variables are only supported in string templates when using
\code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}.
}
